+++ utils/BinaryArchiveWriter.h
@@ -30,7 +30,7 @@ namespace Kernel
 
         virtual bool HasError() override;
         virtual bool IsWriter() override;
-        virtual size_t GetBufferSize() override;
+        virtual uint32_t GetBufferSize() override;
         virtual const char* GetBuffer() override;
 
         size_t capacity;
